# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Gym Tracker Core API
  version: 0.1.0
  description: >
    Single source-of-truth contract between the Core API (apps/api) and every client
    (bot, admin, web, miniapp, future mobile). Covers every data operation the bot
    performs today (so Phase 3 can move the bot off direct SQL) plus the admin operations.


    Authentication: a per-user session JWT issued by the auth endpoints, sent as
    `Authorization: Bearer <token>` (`userJwt`). Per-user scoping is derived from the
    authenticated identity (`sub` claim) — clients never pass `user_id` in the body or query.
    Bot-facing operations also accept a service token (`serviceAuth`, the `X-Service-Token`
    header) together with `X-Act-As-User` (the Telegram id the service acts on behalf of).
servers:
  - url: /api/v1
    description: Core API v1
...
paths:
...
  /training/days:
    get:
      tags: [training]
      summary: List the caller's training days, grouped and reverse-chronological
      description: >
        Day-grouped view of the caller's training history for the History feature:
        one entry per day the caller trained, newest day first. The optional `from`
        and `to` query params bound the window (inclusive, date part only);
        omitting them returns a default window. Dates are naive-tolerant (GYM-30).
      operationId: listTrainingDays
      security:
        - userJwt: []
        - serviceAuth: []
      parameters:
        - $ref: '#/components/parameters/ActAsUser'
        - name: from
          in: query
          required: false
          description: Inclusive start date of the window (date part only).
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          description: Inclusive end date of the window (date part only).
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/TimezoneQuery'
      responses:
        '200':
          description: Training days for the caller, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingDay'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /training/day/{date}:
    get:
      tags: [training]
      summary: Detail of the caller's training on a single day
      description: >
        All sets the caller recorded on the given day, grouped by exercise with
        muscle and exercise names denormalized so the client does not have to join.
        The date is naive-tolerant (GYM-30).
      operationId: getTrainingDay
      security:
        - userJwt: []
        - serviceAuth: []
      parameters:
        - name: date
          in: path
          required: true
          description: Calendar date of the training day (date part only).
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/ActAsUser'
      responses:
        '200':
          description: The caller's training detail for the day.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingDayDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /training/{training_id}:
    put:
      tags: [training]
      summary: Update weight and reps of an existing training record
      description: >
        Maps to `update_training_data(training_id, user_id, weight, reps)`. Only the
        caller's own record can be updated; only weight and reps are mutable.
      operationId: updateTraining
      security:
        - userJwt: []
        - serviceAuth: []
      parameters:
        - $ref: '#/components/parameters/TrainingId'
        - $ref: '#/components/parameters/ActAsUser'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingUpdate'
      responses:
        '200':
          description: The updated training record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [training]
      summary: Delete a training set owned by the authenticated user
      description: >
        Deletes the caller's own recorded set. Only the caller's own record can be
        removed (RLS-scoped per identity / X-Act-As-User).
      operationId: deleteTraining
      security:
        - userJwt: []
        - serviceAuth: []
      parameters:
        - $ref: '#/components/parameters/TrainingId'
        - $ref: '#/components/parameters/ActAsUser'
      responses:
        '204':
          description: Training set deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /training/{training_id}/move:
    patch:
      tags: [training]
      summary: Move a training set to another day and/or exercise
      description: >
        Moves the caller's own set to another calendar day (date) and/or another
        exercise (muscle_name + exercise_name). Only the caller's own record can be
        moved (RLS-scoped per identity / X-Act-As-User). At least one of
        {date, (muscle_name + exercise_name)} must be provided; the set's number,
        weight and reps are preserved. History v2 (GYM-51).
      operationId: moveTrainingSet
      security:
        - userJwt: []
        - serviceAuth: []
      parameters:
        - $ref: '#/components/parameters/TrainingId'
        - $ref: '#/components/parameters/ActAsUser'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingMove'
      responses:
        '200':
          description: The moved training record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Training set not found or not owned by the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            The target day + exercise already has a set with the same set number
            (collision).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: >
            Invalid body, nothing to move (no date and no muscle_name+exercise_name),
            or the target exercise does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------------------------------------------------------------- analytics
...
components:
  securitySchemes:
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Per-user session JWT issued by the /auth endpoints, sent as
        `Authorization: Bearer <token>`. Per-user scoping is derived from the
        `sub` claim. Used by web/miniapp/admin and accepted by the bot-facing
        operations.
    serviceAuth:
      type: apiKey
      in: header
      name: X-Service-Token
      description: >
        Service-to-service token identifying a trusted backend (the bot). Sent
        as the `X-Service-Token` header. A service request MUST also carry the
        companion header `X-Act-As-User` (modeled as the `ActAsUser` parameter
        on each bot-facing operation): the integer Telegram id the service acts
        on behalf of, used for per-user scoping in place of a JWT `sub` claim.
  parameters:
    ActAsUser:
      name: X-Act-As-User
      in: header
      required: false
      description: >
        Telegram id the service acts on behalf of. REQUIRED when authenticating
        with `serviceAuth` (the X-Service-Token header); ignored when
        authenticating with a per-user JWT (scoping comes from the `sub` claim).
      schema:
        type: integer
        format: int64
...
    TrainingId:
      name: training_id
      in: path
      required: true
      schema:
        type: string
...
    TimezoneQuery:
      name: tz
      in: query
      required: false
      description: >
        Optional IANA timezone name (e.g. "Asia/Tbilisi", "Europe/Berlin") used to
        compute day/week boundaries for grouping. When provided, days and weeks are
        grouped in this timezone; when omitted, boundaries are computed in UTC
        (unchanged, backward-compatible behavior). Only meaningful on endpoints that
        group results by day or week.
      schema:
        type: string
        example: Asia/Tbilisi
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found or not accessible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

    # ---- auth
...
    Training:
      type: object
      required: [id, date, user_id, muscle_id, exercise_id, set, weight, reps]
      properties:
        id:
          type: string
          description: Server-assigned record id.
        date:
          type: string
          format: date-time
        user_id:
          type: integer
          format: int64
        muscle_id:
          type: integer
        exercise_id:
          type: integer
        set:
          type: integer
        weight:
          type: number
        reps:
          type: number
...
    TrainingUpdate:
      type: object
      required: [weight, reps]
      properties:
        weight:
          type: number
        reps:
          type: number
    TrainingMove:
      type: object
      description: >
        Move an existing set to another calendar day and/or another exercise.
        All fields are OPTIONAL, but at least one of {date, (muscle_name +
        exercise_name)} must be provided — a body that moves nothing is rejected
        (422). To move to another exercise, send BOTH muscle_name and exercise_name.
        Validation (at-least-one, both names together, target exercise exists) is
        enforced in the API.
      properties:
        date:
          type: string
          format: date
          description: >
            Target calendar day to move the set to (date part only; naive-tolerant
            per GYM-30). Omit to keep the set on its current day.
        muscle_name:
          type: string
          description: >
            Target muscle group name — reference to an existing record, normalized
            only (trim + collapse internal whitespace), not length/char-bound.
            Required together with exercise_name when moving to another exercise.
            See docs/validation.md (Name rules — Create vs lookup).
        exercise_name:
          type: string
          description: >
            Target exercise name — reference to an existing record, normalized only
            (trim + collapse internal whitespace), not length/char-bound. Required
            together with muscle_name when moving to another exercise.
            See docs/validation.md (Name rules — Create vs lookup).
    TrainingDay:
      type: object
      description: One day the caller trained, summarised for the History list.
      required: [date, muscles, exercises_count, sets_count, has_pr]
      properties:
        date:
          type: string
          format: date
        muscles:
          type: array
          items:
            type: string
          description: Names of the muscle groups trained that day.
        exercises_count:
          type: integer
          description: Distinct exercises trained that day.
        sets_count:
          type: integer
          description: Total sets recorded that day.
        has_pr:
          type: boolean
          description: >
            True when the day holds the caller's current all-time max-weight set
            of at least one exercise (GYM-136). Semantic is "current max", not
            "was a PR when logged": the marker reflects today's standing weight
            records, so a later heavier set moves the marker to its own day, and
            a tie marks every tying day.
    TrainingSet:
      type: object
      description: A single recorded set within a training day's exercise.
      required: [training_id, set, weight, reps]
      properties:
        training_id:
          type: string
          description: Server-assigned record id of the set.
        set:
          type: integer
        weight:
          type: number
        reps:
          type: number
    TrainingDayExercise:
      type: object
      description: One exercise trained on a day, with denormalized names and its sets.
      required: [exercise_id, exercise_name, muscle_name, sets]
      properties:
        exercise_id:
          type: integer
        exercise_name:
          type: string
        muscle_name:
          type: string
        sets:
          type: array
          items:
            $ref: '#/components/schemas/TrainingSet'
    TrainingDayDetail:
      type: object
      description: Full detail of the caller's training on a single day.
      required: [date, exercises]
      properties:
        date:
          type: string
          format: date
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/TrainingDayExercise'

    # ---- analytics
...
security:
  - userJwt: []